fixed what I think is a bug (see description)
An error occurs when rbinding temp to full_site because temp is initialized as a matrix populated with characters ("") and is rbinded with fields in full_site that are logical (T/F) instead of characters. I populated logical fields in temp as NA instead of "" so it can be rbinded properly.